+++ plugins/mapper-annotated-text/build.gradle
@@ -11,17 +11,11 @@ apply plugin: 'elasticsearch.internal-cluster-test'
 
 esplugin {
   description 'The Mapper Annotated_text plugin adds support for text fields with markup used to inject annotation tokens into the index.'
-  classname 'org.elasticsearch.plugin.mapper.AnnotatedTextPlugin'
+  classname 'org.elasticsearch.index.mapper.annotatedtext.AnnotatedTextPlugin'
 }
 
 restResources {
   restApi {
     include '_common', 'indices', 'index', 'search'
   }
-}
-
-tasks.named('splitPackagesAudit').configure {
-    // TODO: rename this package to be annotated text specific
-  ignoreClasses 'org.elasticsearch.search.fetch.subphase.highlight.AnnotatedPassageFormatter',
-    'org.elasticsearch.search.fetch.subphase.highlight.AnnotatedTextHighlighter'
 }
